Revert "LPD-51094 How can this ever be null?"
This reverts commit 406d0132c34213d17b82516e4b416901c660613b.

+++ modules/apps/sharing/sharing-test-util/src/main/java/com/liferay/sharing/test/util/BaseSharingTestCase.java
@@ -234,10 +234,10 @@ public abstract class BaseSharingTestCase<T extends ClassedModel> {
 
 	@Test
 	public void testInlinePermissionsToUserGroup() throws Exception {
-		UserGroup userGroup = UserGroupTestUtil.addUserGroup();
-
 		PermissionChecker permissionChecker =
 			PermissionCheckerFactoryUtil.create(_groupUser);
+
+		UserGroup userGroup = UserGroupTestUtil.addUserGroup();
 
 		try (ContextUserReplace contextUserReplace = new ContextUserReplace(
 				_groupUser, permissionChecker)) {
@@ -263,10 +263,12 @@ public abstract class BaseSharingTestCase<T extends ClassedModel> {
 			Assert.assertEquals(1, getModelCount(_group));
 		}
 		finally {
-			_userGroupLocalService.deleteUserUserGroup(
-				_groupUser.getUserId(), userGroup);
+			if (userGroup != null) {
+				_userGroupLocalService.deleteUserUserGroup(
+					_groupUser.getUserId(), userGroup);
 
-			_userGroupLocalService.deleteUserGroup(userGroup);
+				_userGroupLocalService.deleteUserGroup(userGroup);
+			}
 		}
 	}
 
@@ -311,15 +313,19 @@ public abstract class BaseSharingTestCase<T extends ClassedModel> {
 					model.getModelClassName(), "1234", null, null, null));
 		}
 		finally {
-			_userGroupLocalService.deleteUserUserGroup(
-				TestPropsValues.getUserId(), userGroup1);
+			if (userGroup1 != null) {
+				_userGroupLocalService.deleteUserUserGroup(
+					TestPropsValues.getUserId(), userGroup1);
 
-			_userGroupLocalService.deleteUserGroup(userGroup1);
+				_userGroupLocalService.deleteUserGroup(userGroup1);
+			}
 
-			_userGroupLocalService.deleteUserUserGroup(
-				TestPropsValues.getUserId(), userGroup2);
+			if (userGroup2 != null) {
+				_userGroupLocalService.deleteUserUserGroup(
+					TestPropsValues.getUserId(), userGroup2);
 
-			_userGroupLocalService.deleteUserGroup(userGroup2);
+				_userGroupLocalService.deleteUserGroup(userGroup2);
+			}
 		}
 	}
